Fase 5 V0
Introduced SeePlaneDataOnClickForm to display detailed plane data when a plane is clicked in the simulation. Added new getter methods to FlightPlan for accessing ID, speed, and positions. Updated SimulationForm to support plane selection and data viewing, and improved cycle management. Enhanced Main form with a close button and improved error handling in CicleSecurityForm.